Add files via upload
New version (230806):
can be set up as Single- or Multi-Throttle, up to 4 throttles.
After the program is uploaded the first time it is configured as Single Throttle.
It can be changed as follows:
When the throttle is turned on it displays the version and then
"Press Stop for setup" for up to 5 seconds. 
If you do nothing within the 5 seconds it starts with the current setting.
If Stop is pushed within 5 seconds the current 'number of throttle' setting is displayed.
Turn the knob to the desired number of throttles. 
Push the knob to take the new setting and exit setup.
Selecting '1 Throttle' puts the device as Single Throttle.